+++ ObjectHeap.hpp
@@ -25,8 +25,8 @@ ScriptContainer *ObjectHeap_Get(int index);
 ScriptObject *ObjectHeap_GetObject(int index);
 ScriptList *ObjectHeap_GetList(int index);
 bool ObjectHeap_SetObjectMember(int index, const ScriptVariant *key, const ScriptVariant *value);
-void ObjectHeap_SetListMember(int index, size_t indexInList, const ScriptVariant *value);
-bool ObjectHeap_InsertInList(int index, size_t indexInList, const ScriptVariant *value);
+void ObjectHeap_SetListMember(int index, uint32_t indexInList, const ScriptVariant *value);
+bool ObjectHeap_InsertInList(int index, uint32_t indexInList, const ScriptVariant *value);
 void ObjectHeap_ListUnfreed();
 
 // turn a white or black object gray (for garbage collection)
